[XGUARD-3542] task pane upgrade
