+++ src/Http/Controllers/LogController.php
