+++ src/resources/js/enums/componentNames.js
@@ -1,0 +1,6 @@
+
+export const componentNames = {
+    KanbanTaskModal: 'KanbanTaskModal',
+    ViewTaskData: 'ViewTaskData',
+    TaskPane: 'TaskPane',
+};
